move header component to lib
