+++ src/routes/(app)/me/admin/invites/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Header from '../../Header.svelte';
+	import Header from '$lib/components/Header.svelte';
 	import type { PageData } from './$types';
 	import InvitesView from './InvitesView.svelte';
